+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.1] - 04:12:2020.
 
-* TODO: Describe initial release.
+* Adds bugsnag notifier
+* Includes the ability to add user details to notifier